Remove unneeded index.ts reexports
