+++ frontend/src/metabase/querying/filters/components/DatePicker/DateShortcutPicker/index.ts
@@ -1,3 +1,2 @@
 export * from "./DateShortcutPicker";
 export { getShortcutOptions } from "./utils";
-export type { ShortcutOption } from "./types";
