+++ frontend/src/metabase/querying/filters/components/DatePicker/RelativeDatePicker/DateIntervalPicker/index.ts
@@ -1,2 +1,1 @@
 export * from "./DateIntervalPicker";
-export * from "./SimpleDateIntervalPicker";
