+++ frontend/src/metabase/querying/filters/components/DatePicker/SpecificDatePicker/DateRangePicker/index.ts
@@ -1,3 +1,2 @@
 export * from "./DateRangePicker";
-export * from "./SimpleDateRangePicker";
 export * from "./types";
