+++ frontend/src/metabase/querying/filters/components/DatePicker/SpecificDatePicker/SingleDatePicker/index.ts
@@ -1,3 +1,2 @@
 export * from "./SingleDatePicker";
-export * from "./SimpleSingleDatePicker";
 export * from "./types";
