+++ frontend/src/metabase/querying/filters/components/DatePicker/SpecificDatePicker/index.ts
@@ -1,2 +1,1 @@
 export * from "./SpecificDatePicker";
-export * from "./SimpleSpecificDatePicker";
