+++ frontend/src/metabase/querying/filters/components/DatePicker/index.ts
@@ -1,6 +1,3 @@
 export * from "./DatePicker";
-export * from "./SimpleDatePicker";
-export { getShortcutOptions } from "./DateShortcutPicker";
-export type { ShortcutOption } from "./DateShortcutPicker";
 export * from "./types";
 export * from "./utils";
