+++ frontend/src/metabase/querying/filters/components/DatePicker/types.ts
@@ -59,3 +59,9 @@ export type DatePickerValue =
 export type DatePickerValueType = DatePickerValue["type"];
 
 export type RelativeIntervalDirection = "last" | "current" | "next";
+
+export interface ShortcutOption {
+  label: string;
+  shortcut: DatePickerShortcut;
+  value: RelativeDatePickerValue;
+}
